♻️ move timer to @datadog/js-core/util
