+++ packages/browser-core/src/browser/cookieAccess.ts
@@ -1,6 +1,5 @@
 import { ONE_MINUTE, ONE_SECOND, dateNow } from '@datadog/js-core/time'
-import { globalObject, mockable } from '@datadog/js-core/util'
-import { setInterval, clearInterval } from '../tools/timer'
+import { globalObject, mockable, setInterval, clearInterval } from '@datadog/js-core/util'
 import { Observable } from '../tools/observable'
 import { display } from '../tools/display'
 import { generateUUID } from '../tools/utils/stringUtils'
