+++ packages/browser-core/src/domain/session/sessionManager.ts
@@ -9,12 +9,18 @@ import {
   relativeNow,
 } from '@datadog/js-core/time'
 import type { TimeStamp, RelativeTime } from '@datadog/js-core/time'
-import { isWorkerEnvironment, mockable } from '@datadog/js-core/util'
+import {
+  isWorkerEnvironment,
+  mockable,
+  clearInterval,
+  clearTimeout,
+  setInterval,
+  setTimeout,
+} from '@datadog/js-core/util'
 import { monitorError } from '@datadog/js-core/monitor'
 import { Observable } from '../../tools/observable'
 import { createValueHistory } from '../../tools/valueHistory'
 import { addEventListener, addEventListeners, DOM_EVENT } from '../../browser/addEventListener'
-import { clearInterval, clearTimeout, setInterval, setTimeout } from '../../tools/timer'
 import { noop, throttle } from '../../tools/utils/functionUtils'
 import { generateUUID } from '../../tools/utils/stringUtils'
 import type { Configuration } from '../configuration'
