+++ packages/browser-core/src/transport/flushController.ts
@@ -1,10 +1,9 @@
 import { ONE_SECOND } from '@datadog/js-core/time'
 import type { Duration } from '@datadog/js-core/time'
-import { isWorkerEnvironment } from '@datadog/js-core/util'
+import { isWorkerEnvironment, clearTimeout, setTimeout } from '@datadog/js-core/util'
+import type { TimeoutId } from '@datadog/js-core/util'
 import type { PageMayExitEvent, PageExitReason } from '../browser/pageMayExitObservable'
 import { Observable } from '../tools/observable'
-import type { TimeoutId } from '../tools/timer'
-import { clearTimeout, setTimeout } from '../tools/timer'
 import { RECOMMENDED_REQUEST_BYTES_LIMIT } from './httpRequest'
 
 export type UrgentFlushReason = PageExitReason
