+++ packages/browser-core/src/transport/sendWithRetryStrategy.ts
@@ -1,7 +1,6 @@
 import { ONE_MINUTE, ONE_SECOND } from '@datadog/js-core/time'
 import type { TrackType } from '@datadog/js-core/transport'
-import { ONE_MEBI_BYTE, ONE_KIBI_BYTE } from '@datadog/js-core/util'
-import { setTimeout } from '../tools/timer'
+import { ONE_MEBI_BYTE, ONE_KIBI_BYTE, setTimeout } from '@datadog/js-core/util'
 import { isServerError } from '../tools/utils/responseUtils'
 import type { Observable } from '../tools/observable'
 import type { Payload, HttpRequestEvent, HttpResponse, BandwidthStats } from './httpRequest'
