+++ packages/browser-debugger/src/domain/deliveryApi.ts
@@ -1,7 +1,7 @@
-import { globalObject, mockable } from '@datadog/js-core/util'
-import type { TimeoutId } from '@datadog/browser-core'
+import { globalObject, mockable, setInterval, clearInterval } from '@datadog/js-core/util'
+import type { TimeoutId } from '@datadog/js-core/util'
 import { dateNow } from '@datadog/js-core/time'
-import { addTelemetryDebug, fetch, isServerError, setInterval, clearInterval } from '@datadog/browser-core'
+import { addTelemetryDebug, fetch, isServerError } from '@datadog/browser-core'
 import { INTAKE_SITE_US1 } from '@datadog/js-core/transport'
 import type { Site } from '@datadog/js-core/transport'
 import { display } from './display'
