+++ packages/browser-rum-core/src/domain/resource/resourceCollection.ts
@@ -7,9 +7,8 @@ import {
   ResourceType,
   runOnReadyState,
   safeTruncate,
-  setTimeout,
 } from '@datadog/browser-core'
-import { mockable, combine } from '@datadog/js-core/util'
+import { mockable, combine, setTimeout } from '@datadog/js-core/util'
 import type { MatchOption } from '@datadog/browser-core'
 import type { Duration } from '@datadog/js-core/time'
 import { elapsed, relativeToClocks, toServerDuration } from '@datadog/js-core/time'
