+++ packages/browser-rum-core/src/domain/startCustomerDataTelemetry.ts
@@ -1,7 +1,8 @@
 import type { FlushEvent, Observable, Telemetry } from '@datadog/browser-core'
+import { setInterval } from '@datadog/js-core/util'
 import type { Context } from '@datadog/js-core/util'
 import { ONE_SECOND } from '@datadog/js-core/time'
-import { addTelemetryMetrics, setInterval, TelemetryMetrics } from '@datadog/browser-core'
+import { addTelemetryMetrics, TelemetryMetrics } from '@datadog/browser-core'
 import type { LifeCycle } from './lifeCycle'
 import { LifeCycleEventType } from './lifeCycle'
