+++ packages/browser-rum-core/src/domain/waitPageActivityEnd.ts
@@ -1,9 +1,10 @@
-import type { Subscription, TimeoutId } from '@datadog/browser-core'
+import type { Subscription } from '@datadog/browser-core'
 import { timeStampNow } from '@datadog/js-core/time'
 import type { TimeStamp } from '@datadog/js-core/time'
 import { monitor } from '@datadog/js-core/monitor'
-import { matchList, Observable, setTimeout, clearTimeout } from '@datadog/browser-core'
-import { mockable } from '@datadog/js-core/util'
+import { matchList, Observable } from '@datadog/browser-core'
+import type { TimeoutId } from '@datadog/js-core/util'
+import { mockable, setTimeout, clearTimeout } from '@datadog/js-core/util'
 import { createPerformanceObservable, RumPerformanceEntryType } from '../browser/performanceObservable'
 import type { RumMutationRecord } from '../browser/domMutationObservable'
 import { isElementNode } from '../browser/htmlDomUtils'
