+++ packages/browser-rum/src/domain/profiling/types.ts
@@ -1,6 +1,5 @@
-import type { Profiler } from '@datadog/js-core/util'
+import type { Profiler, TimeoutId } from '@datadog/js-core/util'
 import type { ClocksState } from '@datadog/js-core/time'
-import type { TimeoutId } from '@datadog/browser-core'
 import type { BrowserProfileEvent, BrowserProfilerTrace, RumViewEntry } from '../../types'
 import type { LongTaskContext } from './longTaskHistory'
