+++ packages/browser-rum/src/domain/segmentCollection/segmentCollection.ts
@@ -1,6 +1,8 @@
-import type { DeflateEncoder, HttpRequest, TimeoutId, SessionManager } from '@datadog/browser-core'
+import type { DeflateEncoder, HttpRequest, SessionManager } from '@datadog/browser-core'
 import { ONE_SECOND } from '@datadog/js-core/time'
-import { isPageExitReason, PageExitReason, clearTimeout, setTimeout } from '@datadog/browser-core'
+import { isPageExitReason, PageExitReason } from '@datadog/browser-core'
+import type { TimeoutId } from '@datadog/js-core/util'
+import { clearTimeout, setTimeout } from '@datadog/js-core/util'
 import type { LifeCycle, ViewHistory, RumConfiguration } from '@datadog/browser-rum-core'
 import { LifeCycleEventType } from '@datadog/browser-rum-core'
 import type { BrowserRecord, CreationReason, SegmentContext } from '../../types'
