+++ packages/js-core/src/entries/util.ts
@@ -41,3 +41,5 @@ export { jsonStringify, detachToJsonMethod } from '../util/jsonStringify'
 export type { ObjectWithToJsonMethod } from '../util/jsonStringify'
 export { getZoneJsOriginalValue } from '../util/getZoneJsOriginalValue'
 export type { BrowserWindowWithZoneJs } from '../util/getZoneJsOriginalValue'
+export type { TimeoutId } from '../util/timer'
+export { setTimeout, clearTimeout, setInterval, clearInterval } from '../util/timer'
